+++ api/decode.go
@@ -23,7 +23,7 @@ func decodeActivity(request *http.Request) (*models.Activity, *models.Actor, []b
 		return nil, nil, nil, err
 	}
 	KeyID := verifier.KeyId()
-	keyOwnerActor, err := models.NewActivityPubActorFromRemoteActor(KeyID, fmt.Sprintf("%s (golang net/http; Activity-Relay %s; %s)", GlobalConfig.ServerServiceName(), version, GlobalConfig.ServerHostname().Host), ActorCache)
+	keyOwnerActor, err := models.NewActivityPubActorFromRemoteActor(KeyID, RelayActor.ID, fmt.Sprintf("%s (golang net/http; Activity-Relay %s; %s)", GlobalConfig.ServerServiceName(), version, GlobalConfig.ServerHostname().Host), GlobalConfig.ActorKey(), ActorCache)
 	if err != nil {
 		return nil, nil, nil, err
 	}
@@ -56,7 +56,7 @@ func decodeActivity(request *http.Request) (*models.Activity, *models.Actor, []b
 	if err != nil {
 		return nil, nil, nil, err
 	}
-	remoteActor, err := models.NewActivityPubActorFromRemoteActor(activity.Actor, fmt.Sprintf("%s (golang net/http; Activity-Relay %s; %s)", GlobalConfig.ServerServiceName(), version, GlobalConfig.ServerHostname().Host), ActorCache)
+	remoteActor, err := models.NewActivityPubActorFromRemoteActor(activity.Actor, RelayActor.ID, fmt.Sprintf("%s (golang net/http; Activity-Relay %s; %s)", GlobalConfig.ServerServiceName(), version, GlobalConfig.ServerHostname().Host), GlobalConfig.ActorKey(), ActorCache)
 	if err != nil {
 		return nil, nil, nil, err
 	}
@@ -69,7 +69,7 @@ func fetchOriginalActivityFromURL(url string) (*models.Activity, *models.Actor, 
 	if err != nil {
 		return nil, nil, err
 	}
-	remoteActor, err := models.NewActivityPubActorFromRemoteActor(remoteActivity.Actor, fmt.Sprintf("%s (golang net/http; Activity-Relay %s; %s)", GlobalConfig.ServerServiceName(), version, GlobalConfig.ServerHostname().Host), ActorCache)
+	remoteActor, err := models.NewActivityPubActorFromRemoteActor(remoteActivity.Actor, RelayActor.ID, fmt.Sprintf("%s (golang net/http; Activity-Relay %s; %s)", GlobalConfig.ServerServiceName(), version, GlobalConfig.ServerHostname().Host), GlobalConfig.ActorKey(), ActorCache)
 	if err != nil {
 		return &remoteActivity, nil, err
 	}
